[Silver III] Title: 피보나치 함수, Time: 156 ms, Memory: 16028 KB -BaekjoonHub

+++ 백준/Silver/1003. 피보나치 함수/피보나치 함수.java
@@ -2,7 +2,7 @@ import java.io.*;
 
 public class Main {
     public static void main(String[] args) throws IOException{
-        long[][] dp = new long[41][2];
+        int[][] dp = new int[41][2];
         dp[0][0] = 1;
         dp[0][1] = 0;
         dp[1][0] = 0;
